Move str2getdns type functions in the library
